Corrections based on Robert's comments
