+++ main.py
@@ -6,7 +6,7 @@
 # import testcases.testcase_bc_tuning
 
 # Run test case to tune diameters or transmissibilitites with the adjoint method
-# import testcases.testcase_inverse_problem
+import testcases.testcase_inverse_problem
 
 # Run test case with only the blood flow model
-import testcases.testcase_blood_flow_model
+# import testcases.testcase_blood_flow_model
