+++ source/bloodflowmodel/pressure_flow_solver.py
@@ -78,11 +78,10 @@ class PressureFlowSolverPyAMG(PressureFlowSolver):
         :param flownetwork: flow network object
         :type flownetwork: source.flow_network.FlowNetwork
         """
-        A = csr_matrix(flownetwork.system_matrix); b = flownetwork.rhs
-        B = np.ones((A.shape[0], 1), dtype=A.dtype); BH = B.copy()
+        A = csr_matrix(flownetwork.system_matrix)
+        b = flownetwork.rhs
         # Create solver
-        ml = smoothed_aggregation_solver(A, B=B, BH=BH,
-                                         strength=('symmetric', {'theta': 0.0}),
+        ml = smoothed_aggregation_solver(A, strength=('symmetric', {'theta': 0.0}),
                                          smooth=('energy', {'krylov': 'cg', 'maxiter': 2, 'degree': 1, 'weighting': 'local'}),
                                          improve_candidates=[
                                              ('block_gauss_seidel', {'sweep': 'symmetric', 'iterations': 4}), None,
@@ -94,12 +93,18 @@ class PressureFlowSolverPyAMG(PressureFlowSolver):
                                          max_levels=15,
                                          max_coarse=500,
                                          coarse_solver="pinv")
-
         # Solve system
-        tol_solver = np.abs(np.min(flownetwork.system_matrix)) * tol
+        tol_solver = np.abs(np.min(flownetwork.system_matrix)) * tol  # tolerance
+
+        # x0 is the initial guess for the pressure field.
+        # For the blood flow model and the first iteration of the inverse model, we have to initialise the x0 array
+        # based on boundary conditions. Otherwise, in case of the inverse model, x0 should be the previous solution.
+        # Initialisation: In case of only pressure boundary conditions, x0 should be an array with values between
+        # inlet and outlet pressure values. In case of pressure and flow rate boundary conditions, x0 should be an
+        # array with values +/-50% of the pressure boundary value.
         res = []
         if flownetwork.pressure is None:
-            if (1 in flownetwork.boundary_type) and not(2 in flownetwork.boundary_type):  # if there's pressure boundary
+            if (1 in flownetwork.boundary_type) and not(2 in flownetwork.boundary_type):  # only pressure boundaries
                 boundary_inlet = np.max(flownetwork.boundary_val)
                 boundary_outlet = np.min(flownetwork.boundary_val)
                 x0 = boundary_inlet - np.arange(0.001, 1, 0.999/flownetwork.nr_of_vs) * (boundary_inlet - boundary_outlet)
@@ -107,12 +112,17 @@ class PressureFlowSolverPyAMG(PressureFlowSolver):
             elif (1 in flownetwork.boundary_type) and (2 in flownetwork.boundary_type):
                 boundary_pressure_vs = flownetwork.boundary_vs[flownetwork.boundary_type==1]
                 boundary_pressure_val = flownetwork.boundary_val[flownetwork.boundary_type==1]
-                x0 = np.arange(0.001, 1, 0.999/flownetwork.nr_of_vs) * np.max(boundary_pressure_val)
+                x0 = np.arange(0.5, 1.5, 1/flownetwork.nr_of_vs) * np.max(boundary_pressure_val)
                 x0[boundary_pressure_vs] = boundary_pressure_val
             else:
                 sys.exit("Warning Message: only flow boundary conditions were assigned! Define new boundary conditions,"
                          " including at least one pressure boundary condition!")
-            flownetwork.pressure= ml.solve(b, x0=x0, tol=tol_solver, residuals=res, accel="cg", maxiter=600, cycle="V")
+            flownetwork.pressure, info= ml.solve(b, x0=x0, tol=tol_solver, residuals=res, accel="cg", maxiter=600,
+                                                 cycle="V", return_info=True)
         else:
             x0 = flownetwork.pressure
-            flownetwork.pressure = ml.solve(b, x0=x0, tol=tol_solver, residuals=res, accel="cg", maxiter=600, cycle="V")
+            flownetwork.pressure, info = ml.solve(b, x0=x0, tol=tol_solver, residuals=res, accel="cg", maxiter=600,
+                                                  cycle="V", return_info=True)
+        # Provides convergence information
+        if not info == 0:  # if info is zero, successful exit from the iterative solver
+            print("ERROR in Solving the Matrix")
